refactor(analytics): remove deprecated namespaced API
