+++ packages/analytics/typedoc.json
@@ -1,48 +1,5 @@
 {
   "$schema": "https://typedoc.org/schema.json",
-  "entryPoints": ["lib/modular.ts", "lib/types/analytics.ts"],
-  "tsconfig": "tsconfig.json",
-  "intentionallyNotExported": [
-    "_AddPaymentInfoEventParameters",
-    "_AddShippingInfoEventParameters",
-    "_AddShippingInfoParameters",
-    "_AddToCartEventParameters",
-    "_AddToWishlistEventParameters",
-    "_AnalyticsCallOptions",
-    "_AnalyticsSettings",
-    "_BeginCheckoutEventParameters",
-    "_CampaignDetailsEventParameters",
-    "_ConsentSettings",
-    "_EarnVirtualCurrencyEventParameters",
-    "_EventParams",
-    "_GenerateLeadEventParameters",
-    "_GtagConfigParams",
-    "_Item",
-    "_JoinGroupEventParameters",
-    "_LevelEndEventParameters",
-    "_LevelStartEventParameters",
-    "_LevelUpEventParameters",
-    "_LoginEventParameters",
-    "_PostScoreEventParameters",
-    "_PurchaseEventParameters",
-    "_RefundEventParameters",
-    "_RemoveFromCartEventParameters",
-    "_ScreenViewParameters",
-    "_SearchEventParameters",
-    "_SelectContentEventParameters",
-    "_SelectItemEventParameters",
-    "_SelectPromotionEventParameters",
-    "_SetCheckoutOptionEventParameters",
-    "_SettingsOptions",
-    "_ShareEventParameters",
-    "_SignUpEventParameters",
-    "_SpendVirtualCurrencyEventParameters",
-    "_Statics",
-    "_UnlockAchievementEventParameters",
-    "_ViewCartEventParameters",
-    "_ViewItemEventParameters",
-    "_ViewItemListEventParameters",
-    "_ViewPromotionEventParameters",
-    "_ViewSearchResultsParameters"
-  ]
+  "entryPoints": ["lib/index.ts", "lib/types/analytics.ts"],
+  "tsconfig": "tsconfig.json"
 }
