+++ packages/app/lib/common/index.ts
@@ -133,64 +133,6 @@ type InstanceMap = Record<string, MethodMap>;
 type DeprecationMap = Record<string, InstanceMap>;
 
 const mapOfDeprecationReplacements: DeprecationMap = {
-  analytics: {
-    default: {
-      logEvent: 'logEvent()',
-      setAnalyticsCollectionEnabled: 'setAnalyticsCollectionEnabled()',
-      setSessionTimeoutDuration: 'setSessionTimeoutDuration()',
-      getAppInstanceId: 'getAppInstanceId()',
-      getSessionId: 'getSessionId()',
-      setUserId: 'setUserId()',
-      setUserProperty: 'setUserProperty()',
-      setUserProperties: 'setUserProperties()',
-      resetAnalyticsData: 'resetAnalyticsData()',
-      setDefaultEventParameters: 'setDefaultEventParameters()',
-      initiateOnDeviceConversionMeasurementWithEmailAddress:
-        'initiateOnDeviceConversionMeasurementWithEmailAddress()',
-      initiateOnDeviceConversionMeasurementWithHashedEmailAddress:
-        'initiateOnDeviceConversionMeasurementWithHashedEmailAddress()',
-      initiateOnDeviceConversionMeasurementWithPhoneNumber:
-        'initiateOnDeviceConversionMeasurementWithPhoneNumber()',
-      initiateOnDeviceConversionMeasurementWithHashedPhoneNumber:
-        'initiateOnDeviceConversionMeasurementWithHashedPhoneNumber()',
-      setConsent: 'setConsent()',
-      // We're deprecating all helper methods for event. e.g. `logAddPaymentInfo()` from namespaced and modular.
-      logAddPaymentInfo: 'logEvent()',
-      logScreenView: 'logEvent()',
-      logAddShippingInfo: 'logEvent()',
-      logAddToCart: 'logEvent()',
-      logAddToWishlist: 'logEvent()',
-      logAppOpen: 'logEvent()',
-      logBeginCheckout: 'logEvent()',
-      logCampaignDetails: 'logEvent()',
-      logEarnVirtualCurrency: 'logEvent()',
-      logGenerateLead: 'logEvent()',
-      logJoinGroup: 'logEvent()',
-      logLevelEnd: 'logEvent()',
-      logLevelStart: 'logEvent()',
-      logLevelUp: 'logEvent()',
-      logLogin: 'logEvent()',
-      logPostScore: 'logEvent()',
-      logSelectContent: 'logEvent()',
-      logPurchase: 'logEvent()',
-      logRefund: 'logEvent()',
-      logRemoveFromCart: 'logEvent()',
-      logSearch: 'logEvent()',
-      logSelectItem: 'logEvent()',
-      logSetCheckoutOption: 'logEvent()',
-      logSelectPromotion: 'logEvent()',
-      logShare: 'logEvent()',
-      logSignUp: 'logEvent()',
-      logSpendVirtualCurrency: 'logEvent()',
-      logTutorialBegin: 'logEvent()',
-      logTutorialComplete: 'logEvent()',
-      logUnlockAchievement: 'logEvent()',
-      logViewCart: 'logEvent()',
-      logViewItem: 'logEvent()',
-      logViewPromotion: 'logEvent()',
-      logViewSearchResults: 'logEvent()',
-    },
-  },
   auth: {
     default: {
       applyActionCode: 'applyActionCode()',
